# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Bookkeeping Agent API
  description: REST API for the Bookkeeping Agent MVP — transaction management, AI-powered chat, reports, and alerts for small shop owners.
  version: 1.0.0
  contact:
    name: Bookkeeping Agent Team
servers:
  - url: https://api.yourdomain.com/api/v1
    description: Production
  - url: http://localhost:4000/api/v1
    description: Local development
security:
  - BearerAuth: []
...
paths:
...
  /auth/signup:
    post:
      tags: [auth]
      summary: Register a new user
      security: []
      x-rateLimit: '10 requests per hour per IP'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignupRequest'
            example:
              name: Amara Osei
              email: amara@example.com
              phone: '+2348012345678'
              password: 'S3cur3P@ss!'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
  /auth/login:
    post:
      tags: [auth]
      summary: Authenticate and receive tokens
      security: []
      x-rateLimit: '5 attempts per 15 minutes per IP; 15-minute lockout on breach'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              email: amara@example.com
              password: 'S3cur3P@ss!'
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
          headers:
            Set-Cookie:
              description: HttpOnly refresh token cookie (refresh_token=...; HttpOnly; Secure; SameSite=Strict)
              schema:
                type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
  /auth/refresh:
    post:
      tags: [auth]
      summary: Rotate access token using refresh token
      security: []
      description: |
        Reads the refresh token from the HttpOnly `refresh_token` cookie (Path=/api/v1/auth)
        set at signup/login. Every call rotates it: the presented token is revoked and a new
        one is set. Presenting an already-rotated token more than 30 seconds after its
        rotation is treated as theft, and revokes that device's session. The app calls this
        on every load to restore the session, so the response includes the user.
      responses:
        '200':
          description: New access token issued, refresh cookie rotated
          content:
            application/json:
              schema:
                type: object
                required: [accessToken, user]
                properties:
                  accessToken:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /auth/logout:
    post:
      tags: [auth]
      summary: Invalidate refresh token and clear cookie
      security: []
      description: |
        Signs out this device only (other devices stay signed in). Identified by the refresh
        cookie, not the access token, so it works after the access token has expired.
        Always succeeds, even with no cookie.
      responses:
        '204':
          description: Logged out
        '500':
          $ref: '#/components/responses/InternalError'

  # ── SHOPS ────────────────────────────────────
...
  /imports/upload:
    post:
      tags: [imports]
      summary: Upload an Excel or CSV file for processing
      description: Max 10 MB. Returns an importId to use in subsequent steps.
      x-rateLimit: '10 requests per hour'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, shopId]
              properties:
                file:
                  type: string
                  format: binary
                shopId:
                  type: string
                  format: uuid
      responses:
        '202':
          description: File accepted; schema detection in progress
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportJob'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '413':
          description: File too large (max 10 MB)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /imports/{importId}/validate:
    parameters:
      - $ref: '#/components/parameters/importId'
    post:
      tags: [imports]
      summary: Run validation on an uploaded file
      description: Validates all rows and returns per-row errors.
      x-rateLimit: '10 requests per hour'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImportColumnMapping'
      responses:
        '200':
          description: Validation results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportValidationResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /imports/{importId}/preview:
    parameters:
      - $ref: '#/components/parameters/importId'
    get:
      tags: [imports]
      summary: Preview first 20 rows with detected schema
      x-rateLimit: '30 requests per minute'
      responses:
        '200':
          description: Preview data + detected column mappings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportPreview'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /imports/{importId}/confirm:
    parameters:
      - $ref: '#/components/parameters/importId'
    post:
      tags: [imports]
      summary: Confirm and ingest validated rows into the database
      description: Only rows that passed validation are ingested. Error rows are skipped unless `skipErrors` is false.
      x-rateLimit: '5 requests per hour'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                skipErrors:
                  type: boolean
                  default: true
                  description: If false, abort if any row has errors.
      responses:
        '200':
          description: Ingestion summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Import already confirmed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalError'

# ─────────────────────────────────────────────
# COMPONENTS
# ─────────────────────────────────────────────
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    importId:
      name: importId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    BadRequest:
      description: Invalid request body or parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: BAD_REQUEST
            message: 'amount must be a positive number'
            details: []
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: UNAUTHORIZED
            message: 'Token is expired or missing'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: NOT_FOUND
            message: 'Shop not found'
    Conflict:
      description: Conflict with existing resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: CONFLICT
            message: 'Email already registered'
    TooManyRequests:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: TOO_MANY_REQUESTS
            message: 'Too many login attempts. Try again in 15 minutes.'
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds until the client may retry
...
    InternalError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: INTERNAL_SERVER_ERROR
            message: 'An unexpected error occurred'
  schemas:
...
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          description: Machine-readable error code
        message:
          type: string
          description: Human-readable description
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
...
    SignupRequest:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 100
        email:
          type: string
          format: email
        phone:
          type: string
          description: E.164 format e.g. +2348012345678
        password:
          type: string
          minLength: 8
          description: Must contain uppercase, lowercase, digit, special char
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
          description: JWT; expires in 15 minutes
        user:
          $ref: '#/components/schemas/User'
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        alertPreferences:
          $ref: '#/components/schemas/AlertPreferences'
        createdAt:
          type: string
          format: date-time
    AlertPreferences:
      type: object
      properties:
        email:
          type: boolean
          default: true
        sms:
          type: boolean
          default: false
        whatsapp:
          type: boolean
          default: false
        quietHoursStart:
          type: string
          example: '22:00'
        quietHoursEnd:
          type: string
          example: '07:00'
        thresholds:
          type: object
          properties:
            lowCash:
              type: number
              example: 10000
            highPayable:
              type: number
              example: 50000
            overdueReceivableDays:
              type: integer
              example: 30

    # ── Shops ────────────────────────────────────
...
    ImportJob:
      type: object
      properties:
        id:
          type: string
          format: uuid
        shopId:
          type: string
          format: uuid
        filename:
          type: string
        status:
          type: string
          enum: [uploaded, validating, validated, confirmed, failed]
        rowCount:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
    ImportColumnMapping:
      type: object
      description: Maps detected file columns to known transaction fields
      properties:
        date:
          type: string
          description: Column name in file that maps to transaction date
        amount:
          type: string
        description:
          type: string
        type:
          type: string
        counterparty:
          type: string
        category:
          type: string
    ImportValidationResult:
      type: object
      properties:
        importId:
          type: string
          format: uuid
        totalRows:
          type: integer
        validRows:
          type: integer
        errorRows:
          type: integer
        qualityScore:
          type: number
          format: float
          description: 0.0–1.0 data quality score
        errors:
          type: array
          items:
            type: object
            properties:
              row:
                type: integer
              field:
                type: string
              message:
                type: string
    ImportPreview:
      type: object
      properties:
        importId:
          type: string
          format: uuid
        detectedColumns:
          type: array
          items:
            type: object
            properties:
              columnName:
                type: string
              detectedType:
                type: string
                enum: [date, amount, text, unknown]
              suggestedMapping:
                type: string
        previewRows:
          type: array
          maxItems: 20
          items:
            type: object
            additionalProperties: true
    ImportResult:
      type: object
      properties:
        importId:
          type: string
          format: uuid
        rowsIngested:
          type: integer
        rowsSkipped:
          type: integer
        detectedPatterns:
          type: object
          properties:
            topCategory:
              type: string
            estimatedMonthlyRevenue:
              type: number
            estimatedMonthlyExpenses:
              type: number